added more checks and info so that directories would not just be deleted - caused issues where baseline scans would have to be recreated

+++ scripts/redcap/export_mr_sessions_pipeline.py
@@ -64,7 +64,7 @@ def check_file_date(pipeline_file,xnat_file):
 # Returns - True if new files were created, False if not
 #
 
-def export_series( redcap_visit_id, xnat, redcap_key, session_and_scan_list, to_directory, filename_pattern, xnat_dir, verbose=False, timer_label=None):
+def export_series( redcap_visit_id, xnat, redcap_key, session_and_scan_list, to_directory, filename_pattern, xnat_dir, mr_session_report_complete, verbose=False, timer_label=None):
     (subject_label, event_label) = redcap_key
     # List should have at least one "SESSION/SCAN" entry
     if not '/' in session_and_scan_list:
@@ -117,10 +117,20 @@ def export_series( redcap_visit_id, xnat, redcap_key, session_and_scan_list, to_
 
     [ session, scan ] = session_and_scan_list.split( ' ' )[0].split('/')
     if len(pipeline_file_list)  :
-        slog.info(redcap_visit_id  + "_" + scan,"Warning: existing MR images of the pipeline are updated",
+        if mr_session_report_complete > 1: 
+            slog.info(redcap_visit_id  + "_" + scan,"INFO: MRI of pipeline seem outdated",
                       file = to_path_pattern,
                       experiment_xnat_id=session,
-                      session_scan_list = session_and_scan_list )
+                      session_scan_list = session_and_scan_list,
+                      info = "File date of " + pipeline_file_list[0] + " was older than " + xnat_file_search[0] + ". If the files should not be updated simply ignore issue.  If it should be update please set the 'complete status' of MRI Session Report to 'Incomplete' and run script again! " 
+            )
+            return False
+            
+        slog.info(redcap_visit_id  + "_" + scan,"Warning: existing MR images of the pipeline are updated",
+                  file = to_path_pattern,
+                  experiment_xnat_id=session,
+                  session_scan_list = session_and_scan_list )
+                        
         # Remove existing files of that type to make sure we start with clean slate
         for xml_file in pipeline_file_list:
             os.remove(xml_file)
@@ -392,24 +402,36 @@ def copy_manual_pipeline_files( redcap_visit_id, xnat, xnat_eid, to_directory ):
     return files_created
 
 
-def delete_workdir(workdir,redcap_visit_id,verbose=False): 
+def delete_workdir(workdir, redcap_visit_id, mr_session_report_complete, missing_mri,verbose=False): 
     if os.path.exists(workdir):
+        if mr_session_report_complete :
+            slog.info(redcap_visit_id,"Error: Deletion of directory " + workdir + " violates MRI Session Report being in status 'Complete'!",
+                      workdir = workdir,
+                      mri_missing_in_xnat = missing_mri, 
+                      info = "If those MRIs are missing in XNAT, then set the status of the MRI Session Report to 'Incomplete' and run script again with -f option.  If they are not missing, check that the names of the scans in XNAT are correct!")
+            return False
+        
         if verbose :
-            print("Deleting " + workdir)
+            print("Deleting " + workdir +" because " + missing_mri)
         try :
             shutil.rmtree(workdir)            
         except Exception as err_msg:  
             slog.info(redcap_visit_id,"Error: Could not delete directory " + workdir, 
                       err_msg = str(err_msg))
+            return False
+
+    return True 
+        
 
 #
 # Export MR images and associated data to pipeline directory tree.
 #
 # Returns - True if new file as created, False if not
 #
-def export_to_workdir( redcap_visit_id, xnat, session_data, pipeline_workdir, redcap_key, xnat_dir, stroop=(None,None,None), verbose=False, timerFlag=False):
-    new_files_created = False
+def export_to_workdir( redcap_visit_id, xnat, session_data, pipeline_workdir, redcap_key, xnat_dir, mr_session_report_complete,stroop=(None,None,None), verbose=False, timerFlag=False):
 
+    new_files_created = False
+    
     # Export structural data
     pipeline_workdir_structural_main = os.path.join( pipeline_workdir, 'structural');
     pipeline_workdir_structural_native = os.path.join( pipeline_workdir_structural_main, 'native' );
@@ -419,9 +441,9 @@ def export_to_workdir( redcap_visit_id, xnat, session_data, pipeline_workdir, re
         else :
             timerLabel = None
 
-        new_files_created = export_series( redcap_visit_id, xnat, redcap_key, session_data['mri_series_t1'], pipeline_workdir_structural_native, 't1.nii', xnat_dir, verbose=verbose, timer_label= timerLabel ) or new_files_created
+        new_files_created = export_series( redcap_visit_id, xnat, redcap_key, session_data['mri_series_t1'], pipeline_workdir_structural_native, 't1.nii', xnat_dir, mr_session_report_complete, verbose=verbose, timer_label= timerLabel ) 
 
-        new_files_created = export_series( redcap_visit_id, xnat, redcap_key, session_data['mri_series_t2'], pipeline_workdir_structural_native, 't2.nii', xnat_dir, verbose=verbose) or new_files_created
+        new_files_created = export_series( redcap_visit_id, xnat, redcap_key, session_data['mri_series_t2'], pipeline_workdir_structural_native, 't2.nii', xnat_dir, mr_session_report_complete, verbose=verbose) or new_files_created
 
         # Copy ADNI phantom XML file
         if 'NCANDA_E' in session_data['mri_adni_phantom_eid']:
@@ -429,26 +451,40 @@ def export_to_workdir( redcap_visit_id, xnat, session_data, pipeline_workdir, re
             new_files_created = copy_adni_phantom_t1w( redcap_visit_id, xnat, session_data['mri_adni_phantom_eid'], pipeline_workdir_structural_native ) or new_files_created
 
     else :
-        delete_workdir(pipeline_workdir_structural_main,redcap_visit_id,verbose)
+        missing_mri=""
+        if session_data['mri_series_t1'] == '' :
+            missing_mri = "ncanda-t1spgr-v1 "
+        if session_data['mri_series_t2'] == '' :
+            missing_mri+= "ncanda-t2fse-v1"
+        flag = delete_workdir(pipeline_workdir_structural_main,redcap_visit_id,mr_session_report_complete , missing_mri,verbose)
+        if not flag:
+            return new_files_created 
 
     # Export diffusion data
     pipeline_workdir_diffusion_main = os.path.join( pipeline_workdir, 'diffusion' );
     pipeline_workdir_diffusion_native = os.path.join(pipeline_workdir_diffusion_main, 'native' );
     if session_data['mri_series_dti6b500pepolar'] != '' and session_data['mri_series_dti60b1000'] != '':
-        new_files_created = export_series( redcap_visit_id, xnat, redcap_key, session_data['mri_series_dti6b500pepolar'], os.path.join( pipeline_workdir_diffusion_native, 'dti6b500pepolar' ), 'dti6-%n.nii', xnat_dir, verbose=verbose ) or new_files_created
+        new_files_created = export_series( redcap_visit_id, xnat, redcap_key, session_data['mri_series_dti6b500pepolar'], os.path.join( pipeline_workdir_diffusion_native, 'dti6b500pepolar' ), 'dti6-%n.nii', xnat_dir, mr_session_report_complete, verbose=verbose ) or new_files_created
 
-        new_files_created = export_series( redcap_visit_id, xnat, redcap_key, session_data['mri_series_dti60b1000'], os.path.join( pipeline_workdir_diffusion_native, 'dti60b1000' ), 'dti60-%n.nii', xnat_dir, verbose=verbose ) or new_files_created
+        new_files_created = export_series( redcap_visit_id, xnat, redcap_key, session_data['mri_series_dti60b1000'], os.path.join( pipeline_workdir_diffusion_native, 'dti60b1000' ), 'dti60-%n.nii', xnat_dir, mr_session_report_complete, verbose=verbose ) or new_files_created
 
         if session_data['mri_series_dti30b400'] != '' :
-            new_files_created = export_series( redcap_visit_id, xnat, redcap_key, session_data['mri_series_dti30b400'], os.path.join( pipeline_workdir_diffusion_native, 'dti30b400' ), 'dti30-%n.nii', xnat_dir, verbose=verbose ) or new_files_created
+            new_files_created = export_series( redcap_visit_id, xnat, redcap_key, session_data['mri_series_dti30b400'], os.path.join( pipeline_workdir_diffusion_native, 'dti30b400' ), 'dti30-%n.nii', xnat_dir, mr_session_report_complete, verbose=verbose ) or new_files_created
 
         if session_data['mri_series_dti_fieldmap'] != '':
-            new_files_created = export_series( redcap_visit_id, xnat, redcap_key, session_data['mri_series_dti_fieldmap'], os.path.join( pipeline_workdir_diffusion_native, 'fieldmap' ), 'fieldmap-%T%N.nii', xnat_dir, verbose=verbose ) or new_files_created
+            new_files_created = export_series( redcap_visit_id, xnat, redcap_key, session_data['mri_series_dti_fieldmap'], os.path.join( pipeline_workdir_diffusion_native, 'fieldmap' ), 'fieldmap-%T%N.nii', xnat_dir, mr_session_report_complete, verbose=verbose ) or new_files_created
         
 
     else :
-        delete_workdir(pipeline_workdir_diffusion_main,redcap_visit_id,verbose)
-
+        missing_mri=""
+        if session_data['mri_series_dti6b500pepolar'] == '' :
+            missing_mri = "ncanda-dti6b500pepolar-v1 "
+        if session_data['mri_series_dti60b1000'] == '':
+            missing_mri+= "ncanda-dti60b1000-v1"
+        flag = delete_workdir(pipeline_workdir_diffusion_main,redcap_visit_id,mr_session_report_complete , missing_mri,verbose)
+        if not flag:
+            return new_files_created 
+        
     # Export EPI functional data (from DICOM files)
     pipeline_workdir_functional_main = os.path.join( pipeline_workdir, 'restingstate');
     pipeline_workdir_functional_native = os.path.join( pipeline_workdir_functional_main, 'native' );
@@ -458,28 +494,41 @@ def export_to_workdir( redcap_visit_id, xnat, session_data, pipeline_workdir, re
         else :
             timerLabel = None
 
-        new_files_created = export_series( redcap_visit_id, xnat, redcap_key, session_data['mri_series_rsfmri'], os.path.join( pipeline_workdir_functional_native, 'rs-fMRI' ), 'bold-%n.nii', xnat_dir, verbose=verbose, timer_label = timerLabel ) or new_files_created
+        new_files_created = export_series( redcap_visit_id, xnat, redcap_key, session_data['mri_series_rsfmri'], os.path.join( pipeline_workdir_functional_native, 'rs-fMRI' ), 'bold-%n.nii', xnat_dir, mr_session_report_complete,verbose=verbose, timer_label = timerLabel ) or new_files_created
         # Copy rs-fMRI physio files
         new_files_created = copy_rsfmri_physio_files( redcap_visit_id, xnat, session_data['mri_series_rsfmri'], os.path.join( pipeline_workdir_functional_native, 'physio' ) ) or new_files_created
 
-        new_files_created = export_series( redcap_visit_id, xnat, redcap_key, session_data['mri_series_rsfmri_fieldmap'], os.path.join( pipeline_workdir_functional_native, 'fieldmap' ), 'fieldmap-%T%N.nii', xnat_dir, verbose=verbose ) or new_files_created
+        new_files_created = export_series( redcap_visit_id, xnat, redcap_key, session_data['mri_series_rsfmri_fieldmap'], os.path.join( pipeline_workdir_functional_native, 'fieldmap' ), 'fieldmap-%T%N.nii', xnat_dir, mr_session_report_complete, verbose=verbose ) or new_files_created
 
     else :
-        delete_workdir(pipeline_workdir_functional_main,redcap_visit_id,verbose)
+        missing_mri=""
+        if session_data['mri_series_rsfmri'] == '' :
+            missing_mri = "ncanda-rsfmri-v1 "
+        if  session_data['mri_series_rsfmri_fieldmap'] == '':
+            missing_mri+= "ncanda-grefieldmap-v1"
+        flag = delete_workdir(pipeline_workdir_functional_main,redcap_visit_id,mr_session_report_complete , missing_mri,verbose)
+        if not flag:
+            return new_files_created
+        
 
     # Export spiral functional data (from uploaded resources)
     pipeline_workdir_spiralstroop  =  os.path.join( pipeline_workdir, 'spiralstroop' )
     if session_data['mri_eid_spiral_stroop'] != '':
         new_files_created = export_spiral_files( redcap_visit_id, xnat, redcap_key, session_data['mri_eid_spiral_stroop'], pipeline_workdir_spiralstroop, stroop=stroop, verbose=verbose ) or new_files_created
     else :
-        delete_workdir(pipeline_workdir_spiralstroop,redcap_visit_id,verbose)
-
+        missing_mri="mri spiral stroop"
+        flag=delete_workdir(pipeline_workdir_spiralstroop,redcap_visit_id,mr_session_report_complete , missing_mri, verbose)
+        if not flag:
+            return new_files_created
+                            
     pipeline_workdir_spiralrest = os.path.join( pipeline_workdir, 'spiralrest')
     if session_data['mri_eid_spiral_rest'] != '':
         new_files_created = export_spiral_files(redcap_visit_id, xnat, redcap_key, session_data['mri_eid_spiral_rest'], pipeline_workdir_spiralrest, verbose=verbose) or new_files_created
     else :
-        delete_workdir(pipeline_workdir_spiralrest,redcap_visit_id,verbose)
-
+        missing_mri="mri spiral rest"
+        flag = delete_workdir(pipeline_workdir_spiralrest,redcap_visit_id,mr_session_report_complete , missing_mri,verbose)
+        if not flag:
+            return new_files_created
 
     # Copy any manual pipeline override files from all involved experiments
     #   First, extract "Experiment ID" part from each "EID/SCAN" string, unless empty, then make into set of unique IDs.
@@ -496,7 +545,7 @@ def export_to_workdir( redcap_visit_id, xnat, session_data, pipeline_workdir, re
 #
 # Export MR session and run pipeline if so instructed
 #
-def export_and_queue(red2cas, redcap_visit_id, xnat, session_data, redcap_key, pipeline_root_dir, xnat_dir,stroop=(None,None,None), run_pipeline_script=None, verbose=False, timerFlag = False ):
+def export_and_queue(red2cas, redcap_visit_id, xnat, session_data, redcap_key, pipeline_root_dir, xnat_dir,mr_session_report_complete,stroop=(None,None,None), run_pipeline_script=None, verbose=False, timerFlag = False ):
     (subject_label, event_label) = redcap_key
     # Put together pipeline work directory for this subject and visit
     subject_code = session_data['mri_xnat_sid']
@@ -515,7 +564,7 @@ def export_and_queue(red2cas, redcap_visit_id, xnat, session_data, redcap_key, p
     if verbose:
         print(subject_label,'/',subject_code,'/',event_label,'to',pipeline_workdir)
 
-    new_files_created = export_to_workdir(redcap_visit_id,xnat, session_data, pipeline_workdir, redcap_key, xnat_dir, stroop=stroop, verbose=verbose, timerFlag= timerFlag)
+    new_files_created = export_to_workdir(redcap_visit_id,xnat, session_data, pipeline_workdir, redcap_key, xnat_dir, mr_session_report_complete, stroop=stroop, verbose=verbose, timerFlag= timerFlag)
 
     if (new_files_created and run_pipeline_script):
         if verbose:
